Adicionado condição para tratar erros genéricos
