--- src/CommonsModule/exceptionFilter/duplicate-entry.exceptionfilter.ts
@@ -1,20 +1,0 @@
-import { ArgumentsHost, Catch, ExceptionFilter, HttpException } from '@nestjs/common';
-import { QueryFailedError } from 'typeorm';
-import { Request, Response } from 'express';
-
-@Catch(QueryFailedError)
-export class DuplicateEntryExceptionFilter implements ExceptionFilter {
-  catch(exception: HttpException, host: ArgumentsHost) {
-    const ctx = host.switchToHttp();
-    const response: Response = ctx.getResponse<Response>();
-    const request: Request = ctx.getRequest<Request>();
-
-    response
-      .status(409)
-      .json({
-        statusCode: 409,
-        timestamp: new Date().toISOString(),
-        path: request.url,
-      });
-  }
-}
