+++ src/main.ts
@@ -9,7 +9,7 @@ import {
 } from 'typeorm-transactional-cls-hooked';
 import { HttpExceptionFilter } from './CommonsModule/httpFilter/http-exception.filter';
 import 'reflect-metadata';
-import { DuplicateEntryExceptionFilter } from './CommonsModule/exceptionFilter';
+import { QueryFailedErrorExceptionFilter } from './CommonsModule/exceptionFilter';
 
 async function bootstrap() {
   require('dotenv-flow').config();
@@ -35,7 +35,7 @@ async function bootstrap() {
     }),
   );
 
-  app.useGlobalFilters(new DuplicateEntryExceptionFilter());
+  app.useGlobalFilters(new QueryFailedErrorExceptionFilter());
 
   const configService = app.get<ConfigService>(ConfigService);
